Move to feeding the user's LDAP auth token into client calls, we intend to configre the grants for crdl-cache to accept user token as we are intending all access to be set by the users rather than the frontend having the power itself
