+++ conf/application.conf
@@ -43,10 +43,6 @@ microservice {
   }
 }
 
-internal-auth {
-  token = "crdl-cache-token"
-}
-
 play.i18n.langCookieHttpOnly: "true"
 
 # Change this value to true to enable Welsh translations to be loaded from messages.cy, and to display the language toggle
